Move unit tests to new project
