+++ dropbox-sdk-dotnet/Dropbox.Api.Tests/Dropbox.Api.Integration.Tests.csproj
