+++ dropbox-sdk-dotnet/Dropbox.Api/Settings.StyleCop
@@ -1,0 +1,13 @@
+<StyleCopSettings Version="105">
+  <GlobalSettings>
+    <BooleanProperty Name="ViolationsAsErrors">True</BooleanProperty>
+  </GlobalSettings>
+  <Analyzers>
+    <Analyzer AnalyzerId="StyleCop.CSharp.DocumentationRules">
+      <AnalyzerSettings>
+        <StringProperty Name="CompanyName">Dropbox Inc</StringProperty>
+        <StringProperty Name="Copyright">Copyright (c) Dropbox Inc. All rights reserved.</StringProperty>
+      </AnalyzerSettings>
+    </Analyzer>
+  </Analyzers>
+</StyleCopSettings>
